+++ starsky/starsky/clientapp/src/containers/search.tsx
@@ -28,19 +28,18 @@ function Search(archive: IArchiveProps) {
   if (!archive.colorClassUsage) return (<>(Search) => no colorClassUsage</>);
 
   return (<>
-    <MenuSearch/>
+    <MenuSearch />
     <div className={!sidebar ? "archive" : "archive collapsed"}>
-      {sidebar ? <ArchiveSidebar {...archive}/> : ""}
-
+      <ArchiveSidebar {...archive} />
       <div className="content">
         <div className="search-header">
-          <MenuSearchBar defaultText={query}/>
+          <MenuSearchBar defaultText={query} />
         </div>
         <div className="content--header">{archive.collectionsCount ? <>{archive.collectionsCount} resultaten</> : "Geen resultaat"}</div>
-        <SearchPagination {...archive}/>
+        <SearchPagination {...archive} />
         {archive.collectionsCount >= 1 ? <ItemListView {...archive} colorClassUsage={archive.colorClassUsage}> </ItemListView> : null}
         {archive.collectionsCount === 0 ? <div className="folder"><div className="warning-box"> Probeer een andere zoekopdracht</div></div> : null}
-        {archive.fileIndexItems.length >= 20 ? <SearchPagination {...archive}/> : null}
+        {archive.fileIndexItems.length >= 20 ? <SearchPagination {...archive} /> : null}
       </div>
     </div>
   </>