+++ starsky/starsky/clientapp/src/interfaces/IExifStatus.ts
@@ -1,16 +1,3 @@
-// export enum IExifStatus {
-//   Default,
-//   NotFoundNotInIndex,
-//   NotFoundSourceMissing,
-//   NotFoundIsDir,
-//   OperationNotSupported,
-//   DirReadOnly,
-//   ReadOnly,
-//   Unauthorized,
-//   Ok,
-//   Deleted
-// }
-
 export enum IExifStatus {
   Default = "Default" as any,
   ReadOnly = "ReadOnly" as any,
@@ -19,4 +6,5 @@ export enum IExifStatus {
   NotFoundSourceMissing = "NotFoundSourceMissing" as any,
   ServerError = "ServerError" as any,
   IgnoredAlreadyImported = "IgnoredAlreadyImported" as any,
+  FileError = "FileError" as any,
 }