+++ history.md
@@ -16,7 +16,6 @@
 
 ## High priority features planned
 - []  (feature) Creating thumbnails from Web Interface (including status)
-- []  (bugfix) When import a non supported image Ok is shown 
 - [] Change from `Newtonsoft.Json` to `System.Text.Json` __not implemented__
 
 # version 0.1.11 - 2019-12-??
@@ -28,8 +27,13 @@
 - [x]   (bugfix) _Front-end_ _Chrom{e,ium}_ the location after loging in now updated (fixed)
 - [x]   (bugfix) _CORS_ Add AllowCredentials policy for production
 - [x]   (feature) _Front-end_ Next/Prev back in Detailview is now also support when searching
+- [x]   (change) _Front-end_ Remove V1 from main menu
 - [x]   (feature) _API_ Add new endpoint `/api/search/relativeObjects`
 - [x]   (bugfix) _Front-end_ change default outline for Chrome/Safari
+- [x]   (bugfix) _Front-end_ fix document.title undefined error 
+- [x]   (bugfix) _Front-end_ fix issue where on empty searchquery a sidebar is shown
+- [x]   (bugfix) _Front-end_ When import a non supported image Ok is shown 
+- [x]   (feature) _Front-end_ Show GPX Files with a map (powered by leaflet/openstreet maps)
 
 # version 0.1.10 - 2019-12-15
 - [x] (bugfix) Archive => After pressing 'Apply' the updates are not shown